+++ spine-ios/Sources/SpineWrapper/include/spine-cpp-light.h
@@ -1,1 +1,0 @@
-./../../../../spine-cpp/spine-cpp-lite/spine-cpp-light.h