+++ spine-ios/CHANGELOG.md
@@ -6,3 +6,4 @@
 - Basic Mesh Rendering (#2)
 - Spine C++ Swift Wrapper (#3)
 - Load `Atlas` & `SkeletonData` (#4)
+- Generate Swift classes from `spine-cpp-lite.h` (#5)